requirements.txt
# Core Dependencies
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.5.3
Flask-Limiter==3.5.0
gunicorn==21.2.0
python-dotenv==1.0.0

# Database Dependencies
psycopg2-binary==2.9.9
pymongo==4.5.0
redis==5.0.1

# Logging and Monitoring
python-json-logger==2.0.7
Werkzeug==2.3.7

# Testing Dependencies
pytest==7.4.0
pytest-cov==4.1.0
pytest-mock==3.11.1
pytest-asyncio==0.21.1 